fix(Flow): drag overlay moved to sidebar
