+++ packages/lab/src/components/Flow/Sidebar/SidebarGroup/SidebarGroupItem/DraggableSidebarGroupItem.tsx
@@ -1,6 +1,6 @@
-import { useForkRef } from "@mui/material";
-
 import { useRef } from "react";
+
+import { useForkRef } from "@mui/material";
 
 import { useDraggable } from "@dnd-kit/core";
